Fixed issue with service restart on save

+++ sdcard/manu_test/configure_services.sh
@@ -3,27 +3,38 @@
 HACKS_DIR=/mnt/sdcard/hacks
 SERVICES_DIR=/mnt/data/etc/runit
 
-rm -rf ${SERVICES_DIR}
+install_service() {
+	hackName=$1
+	serviceDir=${HACKS_DIR}/${hackName}/meta/service
 
-for hack in ${HACKS_DIR}/* ; do
-	if [ -d ${hack} ]; then
-		hackName=`basename $hack`
-		for serviceDir in ${hack}/meta/service ; do
-			if [ -d ${serviceDir} ]; then
-				cd ${serviceDir}
-				
-				if [ -f ${serviceDir}/.enable ]; then
-					for service in ${serviceDir}/* ; do
-						serviceName=`basename ${service}`
-						hackServiceDir=${SERVICES_DIR}/${hackName}-${serviceName}
-						mkdir -p ${hackServiceDir}
-						cp ${service} ${hackServiceDir}/run
-					done
-				fi
-			fi
-		done;
+	if [ -d ${serviceDir} ] && [ -f ${serviceDir}/.enable ]; then
+		cd ${serviceDir}
+		
+		for service in ${serviceDir}/* ; do
+			serviceName=`basename ${service}`
+			hackServiceDir=${SERVICES_DIR}/${hackName}-${serviceName}
+			mkdir -p ${hackServiceDir}
+			cp ${service} ${hackServiceDir}/run
+		done
 	fi
-done;
+}
 
-export PATH=$PATH:${HACKS_DIR}/runit/bin
-runsvdir ${SERVICES_DIR} & 
+install_services() {
+	rm -rf ${SERVICES_DIR}
+
+	for hack in ${HACKS_DIR}/* ; do
+		if [ -d ${hack} ]; then
+			hackName=`basename $hack`
+			install_service ${hackName}
+		fi
+	done;
+
+	export PATH=$PATH:${HACKS_DIR}/runit/bin
+	runsvdir ${SERVICES_DIR} &
+}
+
+if [ -z $1 ]; then
+	install_services
+else
+	install_service $1
+fi
